java_testgen_app/src/test/java/ChatGPT/Kruskal/AIGenerated/T1/KruskalTest.java
package ChatGPT.Kruskal.AIGenerated.T1;

import org.junit.Before;
import org.junit.Test;

import java.util.Arrays;
import java.util.List;

public class KruskalTest {
    private List<Kruskal.Edge> graph1Edges;
    private List<Kruskal.Edge> graph2Edges;

    @Before
    public void setUp() {
        // Build the edges for Graph 1
        graph1Edges = Arrays.asList(
                new Kruskal.Edge(0, 1, 7),
                new Kruskal.Edge(1, 2, 8),
                new Kruskal.Edge(0, 3, 5),
                new Kruskal.Edge(1, 3, 9),
                new Kruskal.Edge(1, 4, 7),
                new Kruskal.Edge(2, 4, 5),
                new Kruskal.Edge(3, 4, 15),
                new Kruskal.Edge(3, 5, 6),
                new Kruskal.Edge(4, 5, 8),
                new Kruskal.Edge(4, 6, 9),
                new Kruskal.Edge(5, 6, 11)
        );

        // Build the edges for Graph 2
        graph2Edges = Arrays.asList(
                new Kruskal.Edge(0, 1, -3),
                new Kruskal.Edge(0, 4, 1),
                new Kruskal.Edge(4, 1, 4),
                new Kruskal.Edge(2, 1, 9),
                new Kruskal.Edge(2, 4, 3),
                new Kruskal.Edge(4, 3, 2)
        );
    }

    @Test
    public void testKruskalGraph1() {
        List<Kruskal.Edge> mst = Kruskal.runKruskalAlgorithm(graph1Edges, 7);
        // You should compare the obtained minimum spanning tree (mst) to an expected result.
        // Define the expected result manually based on your understanding of the algorithm.
        // Then, assert that the obtained mst is equal to the expected result.
        // For example, you can check if the total weight of the mst is correct.
        // assertEquals(expectedMST, mst);
        // Replace 'expectedMST' with your expected result.
    }

    @Test
    public void testKruskalGraph2() {
        List<Kruskal.Edge> mst = Kruskal.runKruskalAlgorithm(graph2Edges, 5);
        // Define the expected result manually and compare it with the obtained mst.
        // assertEquals(expectedMST, mst);
    }

    // Add more test cases as needed for different scenarios.

}
